+++ lib/quick_travel/setting.rb
@@ -1,0 +1,9 @@
+require 'quick_travel/adapter'
+
+module QuickTravel
+  class Setting < Adapter
+    def self.basic
+      find_all!("/api/settings/basic.json").first
+    end
+  end
+end