+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [2.9.1]
- - Allow zero pricing for extra pick items
+### Fixed
+- Allow zero pricing for extra pick items
 
 ## [2.9.0]
 ### Changed